+++ ChangeLog.md
@@ -11,6 +11,8 @@ General:
 Blob:
 
 - GetBlob on Archive tier blobs now fails as expected.
+- Fixed issue of download 0 size blob with range > 0 should have header "Content-Range: bytes \*/0" in returned error. (issue #2458)
+- Aligned behavior with Azure to ignore invalid range requests for blob downloads. (issue #2458)
 
 Table: